seprate type from interface file, testing with app, fix bug and logic
