+++ src/otp/otp.controller.ts
@@ -1,11 +1,7 @@
 import { Body, Controller, HttpStatus, Inject, Post } from "@nestjs/common";
 import { ApiOperation, ApiResponse } from "@nestjs/swagger";
-import {
-    RequestOtp,
-    RequestOtpResponse,
-    VerifyOtp,
-    IOtpService
-} from "../interfaces";
+import { RequestOtpResponse } from "../types";
+import { RequestOtp, VerifyOtp, IOtpService } from "../interfaces";
 
 @Controller("otp")
 export class OtpController {
