+++ src/services/vendors/CoinStash.ts
@@ -1,7 +1,8 @@
 import { Logger } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
 import { AxiosInstance } from "axios";
-import { ConfigSettings, IVendor, UserSignupRequest } from "../../interfaces";
+import { ConfigSettings } from "../../types";
+import { IVendor, UserSignupRequest } from "../../interfaces";
 
 export class CoinStashVendor implements IVendor {
     name = "CoinStash";
